Wrap most functions into a Perspective class
This is important because the mcapi global functions normally only work
on the first world of a server. If you have more than one world (e.g.:
you are in the Nether, or you use the Multiverse plugins) then they
won't work as expected. With the Perspective class, you can now do:

  x = pov('yourusername')

And you'll have an object whose functions always operate on you and/or
the world you happen to be in right now.